requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
boto3==1.38.23
botocore==1.38.23
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.3
django-celery==3.1.17
django-cors-headers==4.3.1
django-decouple==2.1
django-environ==0.11.2
django-redis==5.4.0
django-restframework==0.0.1
django-storages==1.14.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et_xmlfile==2.0.0
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
kombu==5.4.2
numpy==2.1.3
oauthlib==3.2.2
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.0.0
prompt_toolkit==3.0.48
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
pytz==2024.2
razorpay==1.4.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rest-framework-simplejwt==0.0.2
s3transfer>=0.13.0,<0.14.0
setuptools==75.6.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
wheel==0.45.1
whitenoise==6.9.0
olefile
